bug fixed-some metric added

+++ common/src/main/java/in/nimbo/moama/metrics/Metered.java
@@ -1,5 +1,5 @@
 package in.nimbo.moama.metrics;
 
 public interface Metered {
-    String stat(float delta);
+    String stat(double delta);
 }
